+++ pages/about.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import SessionProvider from '../src/components/session/modules/SessionProvider';
-import { isAuthenticated } from '../utils/authentication/authentication';
-import AboutUsPage from '../src/components/aboutUs/pages/AboutUsPage';
+import SessionProvider from '@components/session/modules/SessionProvider';
+import { isAuthenticated } from '@utils/authentication/authentication';
+import AboutUsPage from '@pages/aboutUs';
 import dynamic from 'next/dynamic';
-import Header from '../src/components/header';
-import { WISHES } from '../utils/constants/search';
+import Header from '@components/header';
+import { WISHES } from '@constants/search';
 
 const TopNavigationBar = dynamic(() => import('../src/components/navbar/modules/TopNavigationBar'), { ssr: false });
 const BottomNavigation = dynamic(() => import('../src/components/navbar/modules/BottomNavigation'), { ssr: false });
@@ -23,7 +23,7 @@ export async function getServerSideProps({ params, req, res, query }) {
 const AboutUs = ({ user }) => {
   return (
     <SessionProvider user={user}>
-      <Header title="About Us | GiftForGood" />
+      <Header title='About Us | GiftForGood' />
       <TopNavigationBar showNews={true} searchDefaultIndex={WISHES} />
       <AboutUsPage />
       <MobileFooter />
