+++ src/pages/aboutUs/components/Section/index.js
